pyproject.toml
[project]
name = "crustdata-search"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
